backend/requirements.txt
flask
nltk
PyPDF2
transformers
torch
openai
python-dotenv
scikit-learn